requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
langchain-core==0.1.12
langchain-community==0.0.12
langchain-google-genai==0.0.3
duckduckgo-search==3.9.6
pyjwt==2.8.0
bcrypt==4.0.1
firebase-admin==6.2.0
gunicorn==21.2.0
